refactored crate::app::assembly into multiple files because it was really cluttered
